Move SerialKind
