+++ runtime/commonMain/src/kotlinx/serialization/Annotations.kt
@@ -7,6 +7,7 @@
 package kotlinx.serialization
 
 import kotlinx.serialization.json.*
+import kotlinx.serialization.descriptors.*
 import kotlin.reflect.*
 
 /**
