+++ runtime/commonMain/src/kotlinx/serialization/internal/CollectionDescriptors.kt
@@ -4,7 +4,6 @@
 
 package kotlinx.serialization.internal
 
-import kotlinx.serialization.*
 import kotlinx.serialization.descriptors.*
 
 internal sealed class ListLikeDescriptor(val elementDescriptor: SerialDescriptor) : SerialDescriptor {
