+++ runtime/commonMain/src/kotlinx/serialization/json/internal/Polymorphic.kt
@@ -5,6 +5,7 @@
 package kotlinx.serialization.json.internal
 
 import kotlinx.serialization.*
+import kotlinx.serialization.descriptors.*
 import kotlinx.serialization.internal.*
 import kotlinx.serialization.json.*
