+++ runtime/commonMain/src/kotlinx/serialization/json/internal/WriteMode.kt
@@ -4,7 +4,6 @@
 
 package kotlinx.serialization.json.internal
 
-import kotlinx.serialization.*
 import kotlinx.serialization.descriptors.*
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.InvalidKeyKindException
